rename catalog -> museum
