+++ app/src/main/kotlin/com/example/ui/MainActivity.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.example.addAddTodo
-import com.example.addCatalog
+import com.example.addMuseum
 import com.example.addDetail
 import com.example.addFavorite
 import com.example.addGrid
@@ -28,11 +28,11 @@ import com.example.addLogIn
 import com.example.addSignUp
 import com.example.addTodo
 import com.example.addUpdateTodo
-import com.example.feature.catalog.canvas.CanvasActivity
-import com.example.feature.catalog.draggablebox.DraggableBoxActivity
-import com.example.feature.catalog.score.ScoreBoardActivity
-import com.example.feature.catalog.spotlight.SpotLightActivity
-import com.example.feature.catalog.threadscard.ThreadsCardActivity
+import com.example.feature.museum.canvas.CanvasActivity
+import com.example.feature.museum.draggablebox.DraggableBoxActivity
+import com.example.feature.museum.score.ScoreBoardActivity
+import com.example.feature.museum.spotlight.SpotLightActivity
+import com.example.feature.museum.threadscard.ThreadsCardActivity
 import com.example.todo.TodoViewModel
 import com.example.ui.utils.compose.CustomBottomNavigationBar
 import com.example.ui.utils.compose.TrainFloatingActionButton
@@ -92,7 +92,7 @@ class MainActivity : ComponentActivity() {
                             addTodo(navController, currentUser)
                             addAddTodo(navController)
                             addUpdateTodo(navController)
-                            addCatalog(
+                            addMuseum(
                                 openScoreBoard = { startActivity(ScoreBoardActivity.createIntent(this@MainActivity)) },
                                 openDraggableBox = { startActivity(DraggableBoxActivity.createIntent(this@MainActivity)) },
                                 openThreadsCard = { startActivity(ThreadsCardActivity.createIntent(this@MainActivity)) },
