+++ app/src/main/kotlin/com/example/ui/utils/compose/CustomBottomNavigationBar.kt
@@ -31,7 +31,7 @@ import com.google.firebase.auth.FirebaseUser
 private val String?.isTop
     get() = this == Routes.Grid.route ||
         this == Routes.Todo.route ||
-        this == Routes.Catalog.route ||
+        this == Routes.Museum.route ||
         this == Routes.LogIn.route ||
         this == Routes.SignUp.route
