+++ feature/catalog/src/main/kotlin/com/example/feature/museum/threadscard/ThreadCard.kt
@@ -1,4 +1,4 @@
-package com.example.feature.catalog.threadscard
+package com.example.feature.museum.threadscard
 
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.FastOutLinearInEasing
